pyproject.toml
[tool.poetry]
name = "doc-indexer"
version = "0.1.0"
description = "CLI tool for indexing documents with ChromaDB vector search"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "doc_indexer", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
chromadb = "^0.4.24"
pypdf = "^4.0.0"
python-pptx = "^0.6.23"
python-docx = "^1.1.0"
click = "^8.1.7"
pydantic = "^2.6.0"
tqdm = "^4.66.0"
numpy = "<2.0"
# LLM providers
langchain = "^0.1.0"
langchain-openai = "^0.0.5"
langchain-community = "^0.0.10"
ollama = "^0.1.7"
# Image processing
pdf2image = "^1.16.3"
pillow = "^10.0.0"
# Configuration
python-dotenv = "^1.0.0"
# Optional OCR
pytesseract = {version = "^0.3.10", optional = true}

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.1.0"
black = "^24.0.0"
ruff = "^0.2.0"
mypy = "^1.8.0"
pytest-mock = "^3.12.0"

[tool.poetry.scripts]
doc-indexer = "doc_indexer.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --cov=src --cov-report=term-missing --cov-fail-under=80"